Backend/requirements.txt
fastapi
uvicorn
dnspython
motor
rdap
python-whois
pydantic
pydantic-settings
